+++ src/common/theme.ts
@@ -1,0 +1,11 @@
+/*
+ * Copyright (c) 2021 Kiyozz.
+ *
+ * All rights reserved.
+ */
+
+export enum Theme {
+  System = 'system',
+  Light = 'light',
+  Dark = 'dark'
+}