+++ src/renderer/hooks/use-theme.ts
@@ -1,0 +1,34 @@
+/*
+ * Copyright (c) 2021 Kiyozz.
+ *
+ * All rights reserved.
+ */
+
+import { useCallback } from 'react'
+
+import { TelemetryEvents } from '../../common/telemetry-events'
+import { Theme } from '../../common/theme'
+import { useApp } from './use-app'
+import { useTelemetry } from './use-telemetry'
+
+export function useTheme(): [Theme, (theme: Theme) => void] {
+  const {
+    setConfig,
+    config: { theme }
+  } = useApp()
+  const { send } = useTelemetry()
+
+  const setTheme = useCallback(
+    (theme: Theme) => {
+      if (![Theme.System, Theme.Light, Theme.Dark].includes(theme)) {
+        return
+      }
+
+      send(TelemetryEvents.SettingsTheme, { theme })
+      setConfig({ theme })
+    },
+    [setConfig, send]
+  )
+
+  return [theme, setTheme]
+}