+++ src/renderer/index.tsx
@@ -9,7 +9,7 @@ import {
   createMemorySource,
   LocationProvider
 } from '@reach/router'
-import { Color, Titlebar } from 'custom-electron-titlebar'
+import { Titlebar } from 'custom-electron-titlebar'
 import debounce from 'debounce-fn'
 import React from 'react'
 import { render } from 'react-dom'
@@ -23,69 +23,41 @@ import { AppProvider } from './hooks/use-app'
 import { CompilationProvider } from './hooks/use-compilation'
 import { FocusProvider } from './hooks/use-focus'
 import { TelemetryProvider } from './hooks/use-telemetry'
+import { TitlebarProvider } from './hooks/use-titlebar'
 import { SettingsProvider } from './pages/settings/settings-context'
 import { Theme } from './theme'
-import { isDark, onDarkPreferenceChanges } from './utils/dark'
+import { darkColor, lightColor } from './utils/color'
+import { isDark } from './utils/dark'
 import { isProduction } from './utils/is-production'
 
 function start() {
-  const darkColor = Color.fromHex('#282425')
-  const darkColorUnfocus = Color.fromHex('#444041')
-  const lightColor = Color.fromHex('#f6f0f1')
-  const lightColorUnfocus = Color.fromHex('#eae5e6')
   const titlebar = new Titlebar({
     backgroundColor: isDark() ? darkColor : lightColor,
     icon: appIcon,
     unfocusEffect: false
   })
 
-  function onBlur(dark: boolean | undefined = undefined) {
-    const usedIsDark = dark ?? isDark()
-    const usedColor = usedIsDark ? darkColorUnfocus : lightColorUnfocus
-    titlebar.updateBackground(usedColor)
-  }
-
-  function onFocus(dark: boolean | undefined = undefined) {
-    const usedIsDark = dark ?? isDark()
-    const usedColor = usedIsDark ? darkColor : lightColor
-    titlebar.updateBackground(usedColor)
-  }
-
-  onDarkPreferenceChanges(matches => {
-    if (document.hasFocus()) {
-      onFocus(matches)
-    } else {
-      onBlur(matches)
-    }
-  })
-
-  window.addEventListener('blur', () => {
-    onBlur()
-  })
-
-  window.addEventListener('focus', () => {
-    onFocus()
-  })
-
   try {
     const history = createHistory(createMemorySource('/'))
 
     render(
-      <Theme>
+      <TitlebarProvider titlebar={titlebar}>
         <LocationProvider history={history}>
           <AppProvider titlebar={titlebar}>
             <CompilationProvider>
               <SettingsProvider>
                 <FocusProvider>
-                  <TelemetryProvider>
-                    <App />
-                  </TelemetryProvider>
+                  <Theme>
+                    <TelemetryProvider>
+                      <App />
+                    </TelemetryProvider>
+                  </Theme>
                 </FocusProvider>
               </SettingsProvider>
             </CompilationProvider>
           </AppProvider>
         </LocationProvider>
-      </Theme>,
+      </TitlebarProvider>,
       document.getElementById('app')
     )
   } catch (e) {