+++ src/renderer/pages/groups/groups-list-item.tsx
@@ -27,7 +27,7 @@ export function GroupsListItem({
     <div className="paper flex items-center">
       <div className="w-full">
         <div className="text-black-800 dark:text-white">{group.name}</div>
-        <div className="text-xs pl-2 text-black-400 dark:text-white">
+        <div className="text-xs pl-2 text-black-600 dark:text-white">
           {!group.isEmpty() ? (
             <>
               {group.scripts